+++ Runtime/Framework/EventBus/IEventBus.cs
@@ -1,51 +1,53 @@
 using UnityEngine.Events;
 
 namespace DT.UniStart {
+  public interface IEvent { }
+
   public interface IEventListener {
-    UnityAction AddListener<T>(UnityAction action);
-    UnityAction<T> AddListener<T>(UnityAction<T> action);
-    UnityAction AddOnceListener<T>(UnityAction action);
-    UnityAction<T> AddOnceListener<T>(UnityAction<T> action);
-    UnityAction RemoveListener<T>(UnityAction action);
-    UnityAction<T> RemoveListener<T>(UnityAction<T> action);
+    UnityAction AddListener<T>(UnityAction action) where T : IEvent;
+    UnityAction<T> AddListener<T>(UnityAction<T> action) where T : IEvent;
+    UnityAction AddOnceListener<T>(UnityAction action) where T : IEvent;
+    UnityAction<T> AddOnceListener<T>(UnityAction<T> action) where T : IEvent;
+    UnityAction RemoveListener<T>(UnityAction action) where T : IEvent;
+    UnityAction<T> RemoveListener<T>(UnityAction<T> action) where T : IEvent;
   }
 
   public static class IEventListenerExtension {
     // echoed AddListener
-    public static UnityAction AddListener<T>(this IEventListener self, out UnityAction named, UnityAction action) {
+    public static UnityAction AddListener<T>(this IEventListener self, out UnityAction named, UnityAction action) where T : IEvent {
       named = action;
       return self.AddListener<T>(action);
     }
-    public static UnityAction<T> AddListener<T>(this IEventListener self, out UnityAction<T> named, UnityAction<T> action) {
+    public static UnityAction<T> AddListener<T>(this IEventListener self, out UnityAction<T> named, UnityAction<T> action) where T : IEvent {
       named = action;
       return self.AddListener(action);
     }
     // echoed RemoveListener
-    public static UnityAction RemoveListener<T>(this IEventListener self, out UnityAction named, UnityAction action) {
+    public static UnityAction RemoveListener<T>(this IEventListener self, out UnityAction named, UnityAction action) where T : IEvent {
       named = action;
       return self.RemoveListener<T>(action);
     }
-    public static UnityAction<T> RemoveListener<T>(this IEventListener self, out UnityAction<T> named, UnityAction<T> action) {
+    public static UnityAction<T> RemoveListener<T>(this IEventListener self, out UnityAction<T> named, UnityAction<T> action) where T : IEvent {
       named = action;
       return self.RemoveListener(action);
     }
     // echoed AddOnceListener
-    public static UnityAction AddOnceListener<T>(this IEventListener self, out UnityAction named, UnityAction action) {
+    public static UnityAction AddOnceListener<T>(this IEventListener self, out UnityAction named, UnityAction action) where T : IEvent {
       named = action;
       return self.AddOnceListener<T>(action);
     }
-    public static UnityAction<T> AddOnceListener<T>(this IEventListener self, out UnityAction<T> named, UnityAction<T> action) {
+    public static UnityAction<T> AddOnceListener<T>(this IEventListener self, out UnityAction<T> named, UnityAction<T> action) where T : IEvent {
       named = action;
       return self.AddOnceListener(action);
     }
   }
 
   public interface IEventInvoker {
-    void Invoke<T>(T e);
+    void Invoke<T>(T e) where T : IEvent;
   }
 
   public static class IEventInvokerExtension {
-    public static void Invoke<T>(this IEventInvoker self) where T : new() => self.Invoke(new T());
+    public static void Invoke<T>(this IEventInvoker self) where T : IEvent, new() => self.Invoke(new T());
   }
 
   public interface IEventBus : IEventListener, IEventInvoker { }